VERSION 0.3.0 - posix_fadvise + CLLed
O_DIRECT is no more the standard,
as I had problems with this on FUSE filesystems.
The new options -a and -d are there for this
(-d only works if -a is absent).

Also posix_fadvise() now is used.

cmpfast now is CLLed, this is, it is Public Domain
as long as it is not covered by any Copyright.
So it is free as in free speech, free beer and
freely born baby. (Babys are not copyrighted, right?)

Also now references the newest tinolib in tino/ (as usual)